refactor(agent): seed export/review/templates Tier-1 barrels
Land the shovel-ready deep-import consolidation the -08-21 readiness pass
recorded as the consensus next step. Add three curated public-surface barrels
(src/agent/{export,review,templates}/index.ts) re-exporting exactly the symbols
the hosts consume, and route the six host import sites through them instead of
deep-reaching leaf modules by path. Shrinks host-agent-import-baseline.json:
cli 11->9, extension 12->11 (five leaf specifiers plus one leaf collapse to
three doors). Behavior-preserving: no runtime logic moved, only the doors hosts
import through; three of the eight planned Tier-1 doors now exist.

typecheck clean; deep-import ratchet green; dead-code ratchet no new findings;
eslint + prettier clean; architecture suite 104/104.
