+++ config/ratchets/host-agent-import-baseline.json
@@ -4,9 +4,7 @@
     "cli": [
       "@agent/core/state/executionRequests",
       "@agent/core/state/TaskState",
-      "@agent/export/chatExportFormatter",
-      "@agent/export/loadChatExportInput",
-      "@agent/export/schemas",
+      "@agent/export",
       "@agent/followUp",
       "@agent/index",
       "@agent/index/platformAgentDirectories",
@@ -32,11 +30,10 @@
       "@agent/implementations/agentCreator/agentCreatorFlow",
       "@agent/index",
       "@agent/index/platformAgentDirectories",
-      "@agent/review/reviewDiff",
-      "@agent/review/reviewIssues",
+      "@agent/review",
       "@agent/runtime",
       "@agent/storage",
-      "@agent/templates/agentTemplateRenderer",
+      "@agent/templates",
       "@agent/trace"
     ],
     "agent": [
