+++ src/agent/export/index.ts
@@ -1,0 +1,21 @@
+/**
+ * Chat export — the cross-host public surface of `src/agent/export`.
+ *
+ * One curated barrel the hosts import instead of deep-reaching each export
+ * module by path: loading a conversation into the export input shape
+ * (`loadChatExportInput`), rendering it to Markdown (`formatChatAsMarkdown`),
+ * and the `ChatExportInput` type hosts name at that seam — decoupling host
+ * code from the export internals' file layout, per the module-level barrel
+ * pattern set by `@agent/runtime` (#10011) and `@agent/followUp`. The R-b
+ * deep-import width ratchet (`config/ratchets/host-agent-import-baseline.json`)
+ * records each host's single `@agent/export` specifier; the former
+ * `@agent/export/{loadChatExportInput,chatExportFormatter,schemas}` deep
+ * imports collapsed to this door.
+ *
+ * Internal export modules keep importing each other by direct path; nothing
+ * inside `src/agent` imports this barrel, so it introduces no import cycle.
+ */
+
+export { loadChatExportInput } from './loadChatExportInput';
+export { formatChatAsMarkdown } from './chatExportFormatter';
+export type { ChatExportInput } from './schemas';
